\name{geneData}
\alias{geneData}
\docType{data}
\title{
Sample Gene Expression Dataset
}
\description{
This is a gene expression dataset simulating information about 499 gene probes and 13 samples, from an Affymetrix U95v2 chip. Data are made up, as well as sample labels. This is adapted from the Biobase-package, version 2.32.0.}
\usage{data("geneData")}
\format{A list of 2, including gene expression data ('$exprs') and the corresponding grouping factor ('$fac').}
\source{Data were adapted from the Biobase package version 2.32.0, and prepared by the J. Ritz Laboratory (S. Chiaretti).}
