% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorhcplot_src.R
\docType{data}
\name{geneData}
\alias{geneData}
\title{Example Gene Expression Dataset}
\format{
A list with 2 elements:
\describe{
  \item{exprs}{A matrix with 499 rows (genes) and 13 columns (samples)
    containing normalized gene expression values.}
  \item{fac}{A factor including the grouping for each sample.}
}
}
\source{
Data were adapted from the Biobase package version 2.32.0, and 
prepared by the J. Ritz Laboratory (S. Chiaretti).
}
\usage{
data("geneData")
}
\description{
This is a gene expression dataset simulating information 
about 499 gene probes and 13 samples, from an Affymetrix U95v2 chip. 
Data are made up, as well as sample labels. 
This dataset is adapted from the Biobase-package, version 2.32.0.
}
\examples{
data(geneData)
print(geneData[[1]][1:10, 1:6])

}
\keyword{datasets}
